requirements/requirements-test.txt
#
# This file is autogenerated by hatch-pip-compile with Python 3.11
#
# [constraints] requirements.txt (SHA256: a6d2573396e4a2a6dabde479fbfff090558e397f290ef7378cef374dcb14a47c)
#
# - pytest
# - pytest-cov
# - vcrpy
# - click
# - lunchable>=1.2.1
# - python-dateutil
# - rich
# - splitwise<3.0.0,>=2.3.0
#

annotated-types==0.6.0
    # via
    #   -c requirements.txt
    #   pydantic
anyio==4.2.0
    # via
    #   -c requirements.txt
    #   httpx
certifi==2023.11.17
    # via
    #   -c requirements.txt
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via
    #   -c requirements.txt
    #   requests
click==8.1.7
    # via
    #   -c requirements.txt
    #   hatch.envs.test
    #   click-plugins
    #   lunchable
click-plugins==1.1.1
    # via
    #   -c requirements.txt
    #   lunchable
coverage==7.4.0
    # via
    #   coverage
    #   pytest-cov
h11==0.14.0
    # via
    #   -c requirements.txt
    #   httpcore
httpcore==1.0.2
    # via
    #   -c requirements.txt
    #   httpx
httpx==0.26.0
    # via
    #   -c requirements.txt
    #   lunchable
idna==3.6
    # via
    #   -c requirements.txt
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==7.0.1
    # via
    #   -c requirements.txt
    #   lunchable
iniconfig==2.0.0
    # via pytest
lunchable==1.2.1
    # via
    #   -c requirements.txt
    #   hatch.envs.test
markdown-it-py==3.0.0
    # via
    #   -c requirements.txt
    #   rich
mdurl==0.1.2
    # via
    #   -c requirements.txt
    #   markdown-it-py
multidict==6.0.4
    # via yarl
oauthlib==3.2.2
    # via
    #   -c requirements.txt
    #   requests-oauthlib
packaging==23.2
    # via pytest
pluggy==1.4.0
    # via pytest
pydantic==2.5.3
    # via
    #   -c requirements.txt
    #   lunchable
pydantic-core==2.14.6
    # via
    #   -c requirements.txt
    #   pydantic
pygments==2.17.2
    # via
    #   -c requirements.txt
    #   rich
pytest==7.4.4
    # via
    #   hatch.envs.test
    #   pytest-cov
pytest-cov==4.1.0
    # via hatch.envs.test
python-dateutil==2.8.2
    # via
    #   -c requirements.txt
    #   hatch.envs.test
pyyaml==6.0.1
    # via vcrpy
requests==2.31.0
    # via
    #   -c requirements.txt
    #   requests-oauthlib
    #   splitwise
requests-oauthlib==1.3.1
    # via
    #   -c requirements.txt
    #   splitwise
rich==13.7.0
    # via
    #   -c requirements.txt
    #   hatch.envs.test
    #   lunchable
six==1.16.0
    # via
    #   -c requirements.txt
    #   python-dateutil
sniffio==1.3.0
    # via
    #   -c requirements.txt
    #   anyio
    #   httpx
splitwise==2.5.0
    # via
    #   -c requirements.txt
    #   hatch.envs.test
typing-extensions==4.9.0
    # via
    #   -c requirements.txt
    #   pydantic
    #   pydantic-core
urllib3==2.1.0
    # via
    #   -c requirements.txt
    #   requests
vcrpy==6.0.1
    # via hatch.envs.test
wrapt==1.16.0
    # via vcrpy
yarl==1.9.4
    # via vcrpy
zipp==3.17.0
    # via
    #   -c requirements.txt
    #   importlib-metadata
